+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Added `Viewpoint.(expects,depends)` for declaring expectations and dependencies
 * Added `Viewpoint.metadata` to query those expectations and dependencies
 * Added `Viewpoint.build` to build viewpoints (with particular contexts)
+* Added Ordering#[] that returns the direction associated to an attribute.
+* Fixed Ordering#+ to avoid duplicate attributes. This operator is now called
+  `merge` (with alias to `+`) and looks like Hash#merge, including an block to
+  arbitrate conflicts.
 
 # 0.13.1 / 2013-08-05